Update structure of photofx to match recommendations
