+++ examples/photofx/core/effect/effect.js
