+++ examples/photofx/core/effect/invert-effect.js
@@ -4,20 +4,18 @@
  (c) Copyright 2012 Motorola Mobility, Inc.  All Rights Reserved.
  </copyright> */
 var Montage = require("montage").Montage;
-var Effect = require("effect/effect").Effect;
+var Effect = require("core/effect/effect").Effect;
 
-exports.SepiaEffect = Montage.create(Effect, {
+exports.InvertEffect = Montage.create(Effect, {
 
     applyEffect: {
         value: function(pixels, pixelCount) {
-            var i = 0,
-                    average;
+            var i;
 
             for (i = 0; i < pixelCount; i += 4) {
-                average = (pixels[i  ] + pixels[i+1] + pixels[i+2])/ 3;
-                pixels[i  ] = average + 10; // red
-                pixels[i+1] = average; // green
-                pixels[i+2] = average; // blue
+                pixels[i] = 255 - pixels[i];
+                pixels[i+1] = 255 - pixels[i+1];
+                pixels[i+2] = 255 - pixels[i+2];
             }
         }
     }
