+++ examples/photofx/core/effect/sepia-effect.js
@@ -4,18 +4,20 @@
  (c) Copyright 2012 Motorola Mobility, Inc.  All Rights Reserved.
  </copyright> */
 var Montage = require("montage").Montage;
-var Effect = require("effect/effect").Effect;
+var Effect = require("core/effect/effect").Effect;
 
-exports.MultiplyEffect = Montage.create(Effect, {
+exports.SepiaEffect = Montage.create(Effect, {
 
     applyEffect: {
-        value: function(pixels, pixelCount, multiplier) {
-            var i = 0;
+        value: function(pixels, pixelCount) {
+            var i = 0,
+                    average;
 
             for (i = 0; i < pixelCount; i += 4) {
-                pixels[i  ] = pixels[i  ] * multiplier; // red
-                pixels[i+1] = pixels[i+1] * multiplier; // green
-                pixels[i+2] = pixels[i+2] * multiplier; // blue
+                average = (pixels[i  ] + pixels[i+1] + pixels[i+2])/ 3;
+                pixels[i  ] = average + 10; // red
+                pixels[i+1] = average; // green
+                pixels[i+2] = average; // blue
             }
         }
     }
