+++ examples/photofx/core/point-monitor.js
