+++ examples/photofx/index.html
@@ -8,7 +8,7 @@
 <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
     <title>PhotoFX</title>
-    <link rel="stylesheet" href="bootstrap.min.css">
+    <link rel="stylesheet" href="assets/style/bootstrap.min.css">
     <script src="../../montage.js" data-package="."></script>
 
     <script type="text/montage-serialization">{
